Add remote imports for modules we migrate to monorepo.
This commit adds new static .html pages which contain the metadata for
go remote imports.

This allows us to use for example podman.io/go/storage in the go.mod.
Go then does HTTP query to podman.io/go/storage followed by the HTTP
query to podman.io/go to find out the real URL for git repository.

If real person opens that URL in a browser, they are redirected to
the source-codes of particular module.

Signed-off-by: Jan Kaluza <jkaluza@redhat.com>